perf(pmf): switch to normal approximation when the error gets small
To do this, we first precompute the smallest n that has the desired error, and do this for all s 0 to ..., stopping when the smallest n is 1. After that we use the normal approximation for every n. Once we have the precomputation table, we either use the approximation or the exact formala depending on n and s.

NOTE: This sucks. Benchmarking this, the pmf is +5455.9% the non-"optimized" version. It also might just get stuck sometimes, or maybe it just takes that long.