refactor: change rating fuction to be zero-sum
The previous rating function was constant-sum, with a rating of 0 = loss, 0.5 = tie, 1.0 = win. With the new rating function, -1 = loss, 0 = tie, 1 = win.

In conjecture with that, the rating is now renamed to the more appropriate payoff, which corresponds to the payoff matrix typical of classical game theory.

Also fixed some small bugs