fix(solver)!: fix bug in normal states payoff calculation
Previously for the payoff evaluation function calc_normal_payoff`, bust had a payoff of 0, when it should be -1. This is now corrected.

This requires additional rewrites to the paper, specifically `Optimal Policy Analysis`, which is where the policies are investigated.  

Closes #4

BREAKING CHANGE: This is a breaking change as it fundamentally alters the optimal actions for normal states.