style(paper): edit paper style
The style is meant to be something of a blend between nature and ieee with a bit of my own flair (blame Tufte).

This comes with the billionth switch between png and svg. The issue is that _still_ the png is getting bitcrushed. The svg files don't have weird font clipping so I use them instead. Whatever.