fix: fix debug underflow
In debug mode, the code would not compile because `n-1` underflowed when n = 0. In practice, release mode is completely correct, because the `n` term is 0, so any u32 + 0 = 0.

I fix this by just doing the simple arithmetic n * (n - 1) => n * n - n. This does mean that the computation will overflow more quickly.In reality, n starts overflowing at 2^16 = 65,536. This is already outside the scope of any reasonable ruleset.